complete website setting

See merge request iam00/way_housing_crm!6

+++ database/migrations/2024_01_14_053804_create_commission_deducted_settings_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('start_amount');
             $table->bigInteger('end_amount')->nullable();
-            $table->integer('deducted')->nullable()->comment('%');
+            $table->float('deducted')->nullable()->comment('%');
             
             $table->tinyInteger('status')->default(1)->comment('1= Active, 0= Inactive');
             $table->unsignedBigInteger('created_by')->nullable();
